% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpr_request.R
\name{mpr_request}
\alias{mpr_request}
\title{Request USDA MPR historical data}
\usage{
mpr_request(
  slugs,
  report_time = format(Sys.Date(), format = "\%m/\%d/\%Y"),
  message = T
)
}
\arguments{
\item{slugs}{slug IDs. It can be a single slug ID, or multiple slug IDs.}

\item{report_time}{time to request. The default time is current day. It can be a time interval.}

\item{message}{a binary indicator for displaying warning messages or not?}
}
\value{
The function returns a list with the requested data.
}
\description{
The primary function in the \code{usdampr} package to request data from the USDA LMP API.
}
\details{
To make valid request, users should provide valid slug IDs. The data returned is saved in a list, while each list saves info for a section.
}
\examples{
\dontrun{
test1 <- mpr_request(slug = 2463, report_time = '01/25/2020:05/25/2020') # Get livestock
test2 <- mpr_request(slug = 2991, report_time = '01/25/2020:04/25/2020') # Get daily price data
test3 <- mpr_request(slug = 3346, report_time = '2018:2019') # Get FMMOS data

test4 <- mpr_request(c(2461, 2463), report_time = '01/25/2020:05/25/2020') # Multiple slug IDs
}
}
