% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugInfo.R
\docType{data}
\name{slugInfo}
\alias{slugInfo}
\title{A data set that records the slug information}
\format{An object of class \code{data.frame} with 119 rows and 8 columns.}
\usage{
data(slugInfo)
}
\description{
A data set that records the slug information
}
\examples{
data(slugInfo)
}
\keyword{datasets}
