% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugInfo.R
\docType{data}
\name{slugInfo}
\alias{slugInfo}
\title{A data set lists all available report information as supplied by USDA-AMS}
\format{An object of class \code{data.frame} with 119 rows and 9 columns.}
\usage{
data(slugInfo)
}
\description{
The listed reports describes slug ID, legacy slug ID, report title, report frequency, market type, office, and section names.
}
\examples{
data(slugInfo)
}
\keyword{datasets}
